Safari-609.1.11.1/LayoutTests/media/media-controller-timeupdate.html

<!DOCTYPE html>
<html>
<head>
    <title>MediaController - timeupdate test</title>
    <script src=video-test.js></script>
    <script src=media-file.js></script>
    <script>
        var video;

        var start = function() {
            findMediaElement();
            
            waitForEvent('canplay', canplay);
            video.src = findMediaFile("video", "content/test");
        };
        
        var canplay = function() {
            waitForEventAndEnd('timeupdate');
            run('video.play');
        };
    </script>
</head>
<body>
    <body onload="start()">
        <video id="video" mediaGroup="group" controls autoplay></video>
    </body>
</body>
</html>
