LayoutTests/media/media-controller-timeupdate-expected.txt

EVENT(canplay)
RUN(video.play)
EVENT(timeupdate)
END OF TEST
